requirements.txt
tensorflow>=2.10.0
opencv-python>=4.6.0
numpy>=1.21.0
requests>=2.28.0
firebase-admin>=6.0.0
python-dotenv>=0.19.0
